requirement.txt
pandas==1.5.3 
boto3 
python-dotenv